Breaking out number_bucket.h, adding hook for also collecting performance data for PyCLIF.
